5 acceptance screencast (#3)
* screencast: devise added

* converted ERB to Slim

* sign in spec successful

* authed user can create question

* user can create answer

* user can view all questions list

* user can view question's answers

* user can register and logout

* Question has author and User can delete his question

* answers controller and question spec refactor

* Answer belongs to user. User can delete answer

* mistakes refactor

+++ app/controllers/questions_controller.rb
@@ -1,12 +1,16 @@
 # frozen_string_literal: true
 
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+
   expose :questions, -> { Question.all }
-  expose :question
+  expose(:question)
+  expose :answer, -> { Answer.new }
 
   def create
+    question.author = current_user
     if question.save
-      redirect_to question_path(question)
+      redirect_to question_path(question), notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -21,8 +25,14 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      flash[:notice] = 'Your question deleted successfully'
+      redirect_to questions_path
+    else
+      flash[:error] = 'Delete is not permitted!'
+      redirect_to question_path(question)
+    end
   end
 
   private
